Rename `util.py` → `codecs/_util.py`
