+++ multiaddr/codecs/_util.py
@@ -1,9 +1,3 @@
-import six
-import struct
-
-from .multiaddr import Multiaddr
-
-
 if hasattr(int, 'from_bytes'):
     def packed_net_bytes_to_int(b):
         """Convert the given big-endian byte-string to an int."""
